Refactored the code and fixed a circular dependency
