+++ .gitignore
@@ -1,3 +1,6 @@
+# Firebase credentials
+key.json
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
