--- key.json
@@ -1,12 +1,0 @@
-{
-  "type": "XXXX",
-  "project_id": "XXXX",
-  "private_key_id": "XXXX",
-  "private_key": "XXXX",
-  "client_email": "XXXX",
-  "client_id": "XXXX",
-  "auth_uri": "XXXX",
-  "token_uri": "XXXX",
-  "auth_provider_x509_cert_url": "XXXX",
-  "client_x509_cert_url": "XXXX"
-}
